Add StorageManager
