+++ src/features/game/log/GameLogConstants.ts
@@ -1,8 +1,6 @@
 import FontAssets from '../assets/FontAssets';
 import { screenCenter, screenSize } from '../commons/CommonConstants';
 import { BitmapFontStyle } from '../commons/CommonTypes';
-import { GamePosition } from '../commons/CommonTypes';
-import { DialogueLine } from '../dialogue/GameDialogueTypes';
 
 export const headerTextStyle: BitmapFontStyle = {
   key: FontAssets.zektonFont.key,
@@ -15,25 +13,6 @@ export const logTextStyle: BitmapFontStyle = {
   size: 25,
   align: Phaser.GameObjects.BitmapText.ALIGN_LEFT
 };
-
-export const LogMocks: DialogueLine[] = [
-  {
-    line: 'Testing...',
-    speakerDetail: {
-      speakerId: 'scottie',
-      expression: '',
-      speakerPosition: GamePosition.Middle
-    }
-  },
-  {
-    line: 'Testing...',
-    speakerDetail: {
-      speakerId: 'pixel',
-      expression: '',
-      speakerPosition: GamePosition.Middle
-    }
-  }
-];
 
 const LogConstants = {
   headerTextConfig: { x: screenSize.x * 0.44, y: screenCenter.y * 0.25, oriX: 0.0, oriY: 0.5 },
